feat: Add authorization to protected APIs (#20)
* feat: Implement streaming video file upload with validation and S3 integration

Implemented video file upload functionality using the MultipartReader and added validation checks for file signature, video length, and video type. Also, introduced a new function that takes a fileStream as an argument. Additionally, created the DisableFormValueModelBindingAttribute class to disable model binding.

Changes made in this commit:

- Utilized MultipartReader to handle streaming video file uploads efficiently.
- Implemented validation checks for file signature, video length, and video type to ensure data integrity and security.
- Introduced a new function UploadFileAsync that accepts a fileStream as an argument to facilitate the upload process in S3Service.cs.
- Created the DisableFormValueModelBindingAttribute class to disable model binding

This commit enhances the application by enabling secure and efficient video file uploads with proper validation and integration with Amazon S3.

* chore: Clean up code by fixing nullable warnings and updating comments following convention, encode filename

* feat: Enhance form-data processing and optimize function parameters for file upload

Implemented additional code to read key-value pairs for additional fields from the form-data sent by the client. Also, optimized the function parameters of the UploadFileAsync method and added null checks to address nullable warnings.

* feat: Add Culinary Tips model

* feat: Update UploadFileAsync method to return request.Key

Modified the UploadFileAsync method to return the request.Key instead of a boolean value. This change improves the method's usability and provides access to the uploaded file's key for further processing.

* feat: Modify field name in CulinaryTips model

* feat: Simultaneously add data to database and upload file to S3

mplemented the functionality to add data to the database and upload a file to S3 simultaneously.

Changes made in this commit:

- Added the "CulinaryTips" string to the "mongoCollection" in order to define the table in MongoDB.
- Updated the "MongoDBService.cs" file by adding the "IMongoCollection<CulinaryTips> CulinaryTips" property to enable interaction with the "CulinaryTips" collection.
- Modified the "UploadCulinaryTips" method in the "ContentController" to include the logic for adding data to the database and mapping it with the model.
- Added a method in the service layer to handle the insertion of data into the database.

This commit enhances the application's functionality by allowing simultaneous data insertion into the database and file upload to S3, providing a streamlined and efficient process.

* fix: Add missing closing curly brace

* feat: Add FileServer class library and move file processing code

Created a new class library named "FileServer" and moved all file processing code to it.

Changes made in this commit:

- Created the "FileServer" class library project.
- Added a project reference from Novaan-Server to FileServer.
- Added the FileSignature NuGet package to the FileService Class Library.
- Updated the relevant code to use the FileSignature library for file signature validation.
- Inject S3Service in Content Controller
- Add Bson Annotation to Id of CulinaryTips Model

This commit improves the project structure by separating the file processing logic into its own class library, promoting modularity and code organization.

* refactor: Change MongoDB collection name to camelCase

* feat: Add API endpoint to check file metadata

* fix: Fix code based on partner review

Made several improvements and fixes to the code based on the partner's review:

Changes made in this commit:

- Added nullable checks to ensure proper handling of nullable fields.
- Removed nullable annotations for all fields in the CulinaryTips model class.
- Refactored the ProcessMultipartRequest method to make it more generic for common use.
- Updated the logic to use stream instead of byte array for uploading files to Amazon S3.

These changes address the issues highlighted in the partner review, improving the code's robustness, readability, and performance.

* feat: Add constant variable for further use

* refactor: Remove unnecessary file size conversion

Removed the unnecessary conversion of the file size value since it is agreed that the file size value sent from the front-end to the back-end is always in bytes.

* refactor: Update file size limit calculation

* refactor: Change logic to determine permitted file extensions for recipe and tips & tricks

* feat: Add authorization to protected APIs